# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhotoShare API - Instagram-like photo sharing platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  creator: "creator",
  consumer: "consumer",
} as const;

export interface User {
  id: number;
  username: string;
  displayName: string;
  role: UserRole;
  avatarUrl?: string | null;
  bio?: string | null;
  createdAt: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  creator: "creator",
  consumer: "consumer",
} as const;

export interface CreateUserBody {
  username: string;
  displayName: string;
  role: CreateUserBodyRole;
  avatarUrl?: string | null;
  bio?: string | null;
}

export interface Image {
  id: number;
  url: string;
  thumbnailUrl?: string | null;
  title: string;
  caption?: string | null;
  location?: string | null;
  tags: string[];
  creatorId: number;
  creatorUsername: string;
  creatorDisplayName: string;
  averageRating?: number | null;
  ratingCount: number;
  commentCount: number;
  createdAt: string;
}

export interface Comment {
  id: number;
  imageId: number;
  userId: number;
  username: string;
  displayName: string;
  avatarUrl?: string | null;
  text: string;
  createdAt: string;
}

export interface ImageDetail {
  id: number;
  url: string;
  thumbnailUrl?: string | null;
  title: string;
  caption?: string | null;
  location?: string | null;
  tags: string[];
  creatorId: number;
  creatorUsername: string;
  creatorDisplayName: string;
  creatorAvatarUrl?: string | null;
  averageRating?: number | null;
  ratingCount: number;
  commentCount: number;
  comments: Comment[];
  createdAt: string;
}

export interface ImagesResponse {
  images: Image[];
  total: number;
}

export interface CreateImageBody {
  url: string;
  thumbnailUrl?: string | null;
  title: string;
  caption?: string | null;
  location?: string | null;
  tags?: string[];
  creatorId: number;
}

export interface CreateCommentBody {
  imageId: number;
  userId: number;
  text: string;
}

export interface RateImageBody {
  imageId: number;
  userId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
}

export interface RatingInfo {
  imageId: number;
  averageRating?: number | null;
  ratingCount: number;
  userRating?: number | null;
}

export interface FeedStats {
  totalImages: number;
  totalUsers: number;
  totalComments: number;
  totalRatings: number;
  topRatedImage?: Image | null;
  mostCommentedImage?: Image | null;
}

export type GetImagesParams = {
  limit?: number;
  offset?: number;
  creatorId?: string;
};

export type SearchImagesParams = {
  q: string;
  limit?: number;
  offset?: number;
};

export type GetTrendingImagesParams = {
  limit?: number;
};