# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhotoShare API - Instagram-like photo sharing platform
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  imageId: number;
  userId: number;
  username: string;
  displayName: string;
  avatarUrl?: string | null;
  text: string;
  createdAt: string;
}