# lib/api-zod/src/generated/types/createImageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhotoShare API - Instagram-like photo sharing platform
 * OpenAPI spec version: 0.1.0
 */

export interface CreateImageBody {
  url: string;
  thumbnailUrl?: string | null;
  title: string;
  caption?: string | null;
  location?: string | null;
  tags?: string[];
  creatorId: number;
}