# lib/api-zod/src/generated/types/feedStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhotoShare API - Instagram-like photo sharing platform
 * OpenAPI spec version: 0.1.0
 */
import type { Image } from "./image";

export interface FeedStats {
  totalImages: number;
  totalUsers: number;
  totalComments: number;
  totalRatings: number;
  topRatedImage?: Image | null;
  mostCommentedImage?: Image | null;
}