# lib/api-zod/src/generated/types/rateImageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhotoShare API - Instagram-like photo sharing platform
 * OpenAPI spec version: 0.1.0
 */

export interface RateImageBody {
  imageId: number;
  userId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
}